@@ -2,7 +2,8 @@ class Beer < ActiveRecord::Base
   include AverageRating
 
   belongs_to :brewery
-  has_many :ratings, :dependent => :destroy
+  has_many :ratings, dependent: :destroy
+  has_many :raters, through: :ratings, source: :user
 
   validates :name, presence: true
 
